Fix docker-compose, Readme and shell

+++ initialize-env.sh
@@ -1,10 +1,10 @@
 #!/bin/bash
-[ -f .exec-once ] || {
+if [ -e .exec-once ]; then
+touch .exec-once
 chains=(arbitrum avalanche bsc celo ethereum optimism polygon)
 
 for chain in ${chains[@]}; do
 cp ./chain/$chain/.env.sample ./chain/$chain/.env
 done
 
-touch .exec-once
-}
+fi
